Replace auth repository to user
