+++ pkg/repository/user/repository_test.go
@@ -38,11 +38,11 @@ func TestCreate(t *testing.T) {
 				Password: "qweqweqweqwe",
 			},
 			mock: func() {
-				mock.ExpectQuery(fmt.Sprintf("INSERT INTO %s", UserTableName)).
+				mock.ExpectQuery(fmt.Sprintf("INSERT INTO %s", TableName)).
 					WithArgs("check@check.com", "qweqweqweqwe").
 					WillReturnRows(sqlxmock.NewRows([]string{"id"}).AddRow(1))
 
-				mock.ExpectQuery(fmt.Sprintf("SELECT id, email FROM %s WHERE", UserTableName)).
+				mock.ExpectQuery(fmt.Sprintf("SELECT id, email FROM %s WHERE", TableName)).
 					WithArgs(1).
 					WillReturnRows(sqlxmock.NewRows([]string{"id", "email"}).AddRow("1", "check@check.com"))
 			},
@@ -57,7 +57,7 @@ func TestCreate(t *testing.T) {
 				Password: "",
 			},
 			mock: func() {
-				mock.ExpectQuery(fmt.Sprintf("INSERT INTO %s", UserTableName)).
+				mock.ExpectQuery(fmt.Sprintf("INSERT INTO %s", TableName)).
 					WithArgs("check@check.com", "").
 					WillReturnRows(mock.NewRows([]string{"id"}))
 			},
@@ -129,7 +129,7 @@ func TestRetrieve(t *testing.T) {
 			name: "Should find user",
 			id:   1,
 			mock: func() {
-				mock.ExpectQuery(fmt.Sprintf("SELECT id, email FROM %s WHERE", UserTableName)).
+				mock.ExpectQuery(fmt.Sprintf("SELECT id, email FROM %s WHERE", TableName)).
 					WithArgs(1).
 					WillReturnRows(sqlxmock.NewRows([]string{"id", "email"}).AddRow("1", "check@check.com"))
 			},
@@ -141,7 +141,7 @@ func TestRetrieve(t *testing.T) {
 			name: "Should not find user",
 			id:   58,
 			mock: func() {
-				mock.ExpectQuery(fmt.Sprintf("SELECT id, email FROM %s WHERE", UserTableName)).
+				mock.ExpectQuery(fmt.Sprintf("SELECT id, email FROM %s WHERE", TableName)).
 					WithArgs(58).
 					WillReturnRows(sqlxmock.NewRows([]string{"id", "email"}))
 			},
@@ -181,6 +181,66 @@ func TestRetrieve(t *testing.T) {
 				if usr.Email != testCase.expectedEmail {
 					t.Errorf("Invalid user email, expected: %s, got: %s\n", testCase.expectedEmail, usr.Email)
 				}
+			}
+		})
+	}
+}
+
+func TestExists(t *testing.T) {
+	db, mock, err := sqlxmock.Newx()
+	if err != nil {
+		t.Fatalf("Unexpected error when opening a stub db connection, error: %s\n", err)
+	}
+
+	cases := []struct {
+		name         string
+		email        string
+		password     string
+		mock         func()
+		expectedId   int64
+		errorPresent bool
+	}{
+		{
+			name:     "User exists",
+			email:    "check@check.com",
+			password: "qweqweqweqwe",
+			mock: func() {
+				mock.ExpectQuery(fmt.Sprintf("SELECT id FROM %s", TableName)).
+					WithArgs("check@check.com", "qweqweqweqwe").
+					WillReturnRows(sqlxmock.NewRows([]string{"id"}).AddRow(1))
+			},
+			expectedId:   1,
+			errorPresent: false,
+		},
+		{
+			name:     "User doesn't exists",
+			email:    "check@check.com",
+			password: "qweqweqweqwe",
+			mock: func() {
+				mock.ExpectQuery(fmt.Sprintf("SELECT id FROM %s", TableName)).
+					WithArgs("check@check.com", "qweqweqweqwe").
+					WillReturnRows(sqlxmock.NewRows([]string{"id"}))
+			},
+			expectedId:   0,
+			errorPresent: true,
+		},
+	}
+
+	for _, testCase := range cases {
+		t.Run(testCase.name, func(t *testing.T) {
+			testCase.mock()
+			repo := NewRepository(db)
+			id, err := repo.Exists(context.Background(), testCase.email, testCase.password)
+			if err != nil && !testCase.errorPresent {
+				t.Errorf("Unexpected error, error: %s\n", err.Error())
+			}
+
+			if err == nil && testCase.errorPresent {
+				t.Errorf("Should be error\n")
+			}
+
+			if id != testCase.expectedId {
+				t.Errorf("Invalid id, expected: %d, got: %d\n", testCase.expectedId, id)
 			}
 		})
 	}
